added croosBack component and navigationHelpet
